Latest working, several things added
Now will scan each image and finds a few things. Fills the result
divs with the results, and you can click them to highlight them
on the images.

Next: Figure out why I'm missing several words. Perhaps add a list
of rejected words you can click to add to accepted?
- Stack the images so it's a single page
- Add the ability to add a item by dragging a rectangle
and typing what it is in a popup.